ATen/Dlpack functions -> aten/utils*h
